+++ packages/caelus/src/index.ts
@@ -17,3 +17,4 @@ export * from "./features.js";
 export * from "./compiler.js";
 export * from "./lots.js";
 export * from "./profections.js";
+export * from "./firdaria.js";